removed webpack in favor of vite
