+++ inst/gallery-assets/public/rheroicons.css
@@ -1,0 +1,1 @@
+html,body{padding:0;margin:0;font-family:Arial,Helvetica,sans-serif;font-size:16pt}nav,main,section{padding:0;margin:0}h1,h2,h3{padding:0;margin:0 0 12px;line-height:1.4}p{padding:0;margin:12px 0;line-height:1.6}a{color:currentColor;text-decoration:none;padding-bottom:3px;box-shadow:inset 0 -2px 0 0 currentColor}button{display:block;width:100%;text-emphasis:center;margin:0;font-size:16pt;background-color:transparent;padding:6px 0;border-radius:6px;border:2px solid transparent;cursor:pointer}form{padding:0;margin:0}form legend{font-size:18pt}form fieldset{border:none}.visually-hidden{position:absolute;width:1px;height:1px;margin:-1px;clip:rect(1px,1px,1px,1px);clip:rect(1px 1px 1px 1px);overflow:hidden;white-space:nowrap}.navbar{position:fixed;top:0;left:0;width:100%;background-color:#0b4f71;color:#ebf8ff;z-index:3}.navbar .menu{width:100%;display:grid;grid-template-columns:2fr 2fr;justify-content:center;align-items:center;margin:0;padding:0;list-style:none}.navbar .menu .menu-item{padding:0 32px}.navbar .menu .menu-item p{font-size:14pt;font-weight:700}.navbar .menu .menu-item p .rheroicons{position:relative;top:2px;width:18px;height:18px}.navbar .menu .menu-item:last-child{text-align:right}.navbar .menu .menu-item:last-child svg{width:24px;height:24px}.navbar .menu .menu-item:last-child svg path{fill:#ebf8ff}.main{margin-top:53px;width:100%;min-height:100vh;background-color:#e4e7eb}.main .header{position:relative;display:block;background-color:#0b4f71;color:#f5f7fa;width:100%;padding:2em;box-sizing:border-box;overflow:hidden;z-index:1}.main .header h1{font-size:42pt}.main .header h1 .rheroicons{width:32px;height:32px}.main .header cite{font-size:13pt;text-transform:uppercase;letter-spacing:2px;font-weight:700;font-style:normal}.main .header p{color:#ebf8ff}.main .header:after{content:"";display:block;position:absolute;width:125%;height:400px;left:200px;bottom:-150px;background-color:#2d83ae;transform:rotate(-12deg);transform-origin:center;border-radius:30% 19%;z-index:-1}.main .section{width:90%;margin:0 auto;padding:32px 0;z-index:1}.main .section h1{font-size:32pt}.main .icon-section{padding:32px 0}.main .icon-section h2,.main .icon-section p{display:block;width:90%;margin:0 auto}.main .icon-list{padding:0;list-style:none;width:100%;margin:0 auto;display:grid;grid-template-columns:1fr;gap:14px}.main .icon-list .icon-item{border:1px solid transparent}.main .icon-list .icon-item button{box-sizing:border-box;padding:.8em 1em;font-size:13pt;background-color:#fff}.main .icon-list .icon-item button .rheroicons{display:block;margin:0 auto 12px}.main .icon-list .icon-item button span{display:block}.main .icon-list .icon-item button:hover,.main .icon-list .icon-item button:focus{border-color:#3994c1;background-color:#d1eefc}@media (min-width: 512px){.main .icon-list{grid-template-columns:repeat(2,1fr)}}@media (min-width: 812px){.main .icon-list{grid-template-columns:repeat(3,1fr)}}@media (min-width: 1272px){.main .header{padding:5em 10em}.main .section{max-width:1272px;text-align:center}.main .section #iconSet{margin:14px auto}.main .icon-section h2,.main .icon-section p{text-align:center}.main .icon-list{grid-template-columns:repeat(4,1fr)}}.select-input-group{width:100%;position:relative;margin:12px 0;padding:0}.select-input-group .select-input-parent{display:flex;justify-content:flex-start;align-items:center;background-color:#fff;padding:0;margin-top:12px;border-radius:0}.select-input-group .select-input-parent .select-input-selected{display:inline-block;margin:0;font-size:inherit;line-height:1;overflow:hidden;text-overflow:ellipsis;text-align:left;padding:9px 0 9px 72px;width:80%}.select-input-group .select-input-parent .select-input-parent-icon{display:inline-block;width:20%;transform:rotate(180deg);transform-origin:center;transition:all .2s ease-in-out}.select-input-group .select-input-options{position:absolute;list-style:none;width:100%;top:100%;left:0;padding:0;margin:0;background-color:#fff;z-index:1;box-shadow:0 12px 9px 3px #0000004d}.select-input-group .select-input-options .select-input-option{border-bottom:1px solid #9AA5B1}.select-input-group .select-input-options .select-input-option .select-input-option-button{border-radius:0;text-align:left;padding-left:42px}.select-input-group .select-input-options .select-input-option .select-input-option-button:hover,.select-input-group .select-input-options .select-input-option .select-input-option-button:focus{background-color:#2d83ae;color:#ebf8ff}.select-input-group .select-input-options .select-input-option .select-input-option-button .selected-icon{opacity:0}.select-input-group .select-input-options .select-input-option .select-input-option-button.selected .selected-icon{position:relative;width:18px;height:18px;top:0;left:-6px;opacity:1}.select-input-group .select-input-options .select-input-option:first-child{border-top:1px solid #9AA5B1}.select-input-group .select-input-options .select-input-option:last-child{border-bottom:none}.select-input-group.hidden .select-input-options{position:absolute;width:1px;height:1px;clip:rect(1px,1px,1px,1px);overflow:hidden;white-space:nowrap;box-shadow:none}.select-input-group.hidden .select-input-parent .select-input-parent-icon{transform:rotate(0)}.status-box{display:flex;justify-content:flex-start;align-items:center;position:fixed;bottom:1.5em;margin:0 auto;border-radius:6px;transform:translate(-110%);transition:all .4s ease-in-out;background-color:#fff;box-shadow:0 0 6px 4px #0000004d}.status-box .rheroicons{padding:12px;border-radius:6px 0 0 6px}.status-box .rheroicons.rheroicons-hand-thumb-up{background-color:#05400a;stroke:#f5f7fa}.status-box .rheroicons.rheroicons-exclamation-circle{background-color:#610404;stroke:#f5f7fa}.status-box p{padding:0 12px;line-height:1;margin:0}.status-box.show{transform:translate(1.5em)}#iconSet{width:325px}
