--- inst/gallery-assets/src/scss/_main.scss
@@ -1,151 +1,0 @@
-/*  main */
-.main {
-    margin-top: 53px;
-    width: 100%;
-    min-height: 100vh;
-    background-color: $cool-grey-100;
-
-    .header {
-        position: relative;
-        display: block;
-        background-color: $light-blue-900;
-        color: hsl(216, 33%, 97%);
-        width: 100%;
-        padding: 2em;
-        box-sizing: border-box;
-        overflow: hidden;
-        z-index: 1;
-
-        h1 {
-            font-size: 42pt;
-
-            .rheroicons {
-                width: 32px;
-                height: 32px;
-            }
-        }
-
-        cite {
-            font-size: 13pt;
-            text-transform: uppercase;
-            letter-spacing: 2px;
-            font-weight: bold;
-            font-style: normal;
-        }
-
-        p {
-            color: $light-blue-050;
-        }
-
-        &::after {
-            content: "";
-            display: block;
-            position: absolute;
-            width: 125%;
-            height: 400px;
-            left: 200px;
-            bottom: -150px;
-            background-color: $light-blue-600;
-            transform: rotate(-12deg);
-            transform-origin: center;
-            border-radius: 30% 19%;
-            z-index: -1;
-        }
-    }
-
-    .section {
-        width: 90%;
-        margin: 0 auto;
-        padding: 32px 0;
-        z-index: 1;
-
-        h1 {
-            font-size: 32pt;
-        }
-    }
-
-    .icon-section {
-        padding: 32px 0;
-
-        h2,
-        p {
-            display: block;
-            width: 90%;
-            margin: 0 auto;
-        }
-    }
-
-    .icon-list {
-        padding: 0;
-        list-style: none;
-        width: 100%;
-        margin: 0 auto;
-        display: grid;
-        grid-template-columns: 1fr;
-
-        .icon-item {
-            margin: 12px;
-            border: 1px solid transparent;
-
-            button {
-                padding: 12px;
-                background-color: #ffffff;
-
-                .rheroicons {
-                    display: block;
-                    margin: 0 auto;
-                    margin-bottom: 12px;
-                }
-                span {
-                    display: block;
-                }
-
-                &:hover,
-                &:focus {
-                    border-color: $light-blue-500;
-                    background-color: $light-blue-100;
-                }
-            }
-        }
-    }
-}
-
-
-@media (min-width: 512px) {
-    .main {
-        .icon-list {
-            grid-template-columns: repeat(2, 1fr);
-        }
-    }
-}
-
-@media (min-width: 812px) {
-    .main {
-        .icon-list {
-            grid-template-columns: repeat(3, 1fr);
-        }
-    }
-}
-
-$breakpoint: 972px;
-@media (min-width: $breakpoint) {
-    .main {
-        .header {
-            padding: 5em 10em;
-        }
-
-        .section {
-            max-width: $breakpoint;
-        }
-    
-        .icon-section {
-            h2,
-            p {
-                text-align: center;
-            }
-        }
-        .icon-list {
-            grid-template-columns: repeat(4, 1fr);
-        }
-    }
-}
