--- inst/gallery-assets/src/scss/_navbar.scss
@@ -1,51 +1,0 @@
-// navigation bar
-.navbar {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    background-color: $light-blue-900;
-    // background: linear-gradient(25deg, $light-blue-900 75%, $light-blue-600 50%);
-    color: $light-blue-050;
-    z-index: 3;
-
-    .menu {
-        width: 100%;
-        display: grid;
-        grid-template-columns: 2fr 2fr;
-        justify-content: center;
-        align-items: center;
-        margin: 0;
-        padding: 0;
-        list-style: none;
-
-        .menu-item {
-            padding: 0 32px;
-
-            p {
-                font-size: 14pt;
-                font-weight: bold;
-
-                .rheroicons {
-                    position: relative;
-                    top: 2px;
-                    width: 18px;
-                    height: 18px;
-                }
-            }
-
-            &:last-child {
-                text-align: right;
-                
-                svg {
-                    width: 24px;
-                    height: 24px; 
-
-                    path {
-                        fill: $light-blue-050;
-                    }
-                }
-            }
-        }
-    }
-}
