--- inst/gallery-assets/src/scss/_select_input.scss
@@ -1,112 +1,0 @@
-.select-input-group {
-    width: 100%;
-    position: relative;
-    margin: 12px 0;
-    padding: 0;
-    
-    // styling for parent button
-    .select-input-parent {
-        display: flex;
-        justify-content: flex-start;
-        align-items: center;
-        background-color: #ffffff;
-        padding: 0;
-        margin-top: 12px;
-        border-radius: 0;
-
-        .select-input-selected {
-            display: inline-block;
-            padding: 9px 0;
-            margin: 0;
-            font-size: inherit;
-            line-height: 1;
-            overflow: hidden;
-            text-overflow: ellipsis;
-            text-align: left;
-            padding-left: 72px;
-            width: 80%;
-        }
-        
-        .select-input-parent-icon {
-            display: inline-block;
-            width: 20%;
-            transform: rotate(180deg);
-            transform-origin: center;
-            transition: all 0.2s ease-in-out;
-        }
-    }
-
-    // styling for input buttons
-    .select-input-options {
-        position: absolute;
-        list-style: none;
-        width: 100%;
-        top: 100%;
-        left: 0;
-        padding: 0;
-        margin: 0;
-        background-color: #ffffff;
-        z-index: 1;
-        box-shadow: 0 12px 9px 3px hsla(0, 0%, 0%, 0.3);
-
-        .select-input-option {
-            border-bottom: 1px solid $cool-grey-300;
-
-            .select-input-option-button {
-                border-radius: 0;
-                text-align: left;
-                padding-left: 42px;
-                
-                &:hover, &:focus {
-                    background-color: $light-blue-600;
-                    color: $light-blue-050;
-                }
-
-                .selected-icon {
-                    opacity: 0;
-                }
-
-                &.selected {
-                    .selected-icon {
-                        position: relative;
-                        width: 18px;
-                        height: 18px;
-                        top: 0;
-                        left: -6px;
-                        opacity: 1;
-                    }
-                }
-            }
-
-            &:first-child {
-                border-top: 1px solid $cool-grey-300;
-            }
-
-            &:last-child {
-                border-bottom: none;
-            }
-        }
-    }
-
-    // when hidden
-    &.hidden {
-
-        // hide list
-        .select-input-options {
-            position: absolute;
-            width: 1px;
-            height: 1px;
-            clip: rect(1px, 1px, 1px, 1px);
-            overflow: hidden;
-            white-space: nowrap;
-            box-shadow: none;
-        }
-
-        // unrotate icon
-        .select-input-parent {
-            .select-input-parent-icon {
-                transform: rotate(0deg);
-            }
-        }
-    }
-}
