--- inst/gallery-assets/src/scss/_statuses.scss
@@ -1,41 +1,0 @@
-// status boxes
-.status-box {
-    display: flex;
-    justify-content: flex-start;
-    align-items: center;
-    position: fixed;
-    bottom: 1.5em;
-    margin: 0 auto;
-    border-radius: 6px;
-    transform: translateX(-110%);
-    transition: all 0.4s ease-in-out;
-    background-color: #ffffff;
-    box-shadow: 0 0 6px 4px hsl(0, 0%, 0%, 0.3);
-
-    .rheroicons {
-        width: 25px;
-        height: 25px;
-        padding: 12px;
-        border-radius: 6px 0 0 6px;
-
-        &.rheroicons_thumb_up {
-            background-color: $green-900;
-            stroke: $cool-grey-050;
-        }
-
-        &.rheroicons_exclamation {
-            background-color: $red-900;
-            stroke: $cool-grey-050;
-        }
-    }
-
-    p {
-        padding: 0 12px;
-        line-height: 1;
-        margin: 0;
-    }
-
-    &.show {
-        transform: translateX(1.5em);
-    }
-}
