--- inst/gallery-assets/src/scss/index.scss
@@ -1,19 +1,0 @@
-@import "_palette_20";
-@import "_base",  "_navbar", "_main", "_select_input", "_statuses";
-
-/* visually hidden content */
-.visually-hidden {
-    position: absolute;
-    width: 1px;
-    height: 1px;
-    margin: -1px;
-    clip: rect(1px, 1px, 1px, 1px);
-    clip: rect(1px 1px 1px 1px);
-    overflow: hidden;
-    white-space: nowrap;
-}
-
-
-#iconSet {
-    width: 325px;
-}
