+++ src/index.html
@@ -5,11 +5,9 @@
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Vite App</title>
-    <script type="module" crossorigin src="/rheroicons.js"></script>
-    <link rel="stylesheet" href="/rheroicons.css">
   </head>
   <body>
     <div id="app"></div>
-    
+    <script type="module" src="/index.js"></script>
   </body>
 </html>
