+++ src/scss/base.scss
@@ -1,0 +1,57 @@
+html, body {
+  padding: 0;
+  margin: 0;
+  font-family: Arial, Helvetica, sans-serif;
+  font-size: 16pt;
+}
+
+nav, main, section {
+  padding: 0;
+  margin: 0;
+}
+
+h1, h2, h3 {
+  padding: 0;
+  margin: 0;
+  margin-bottom: 12px;
+  line-height: 1.4;
+}
+
+p {
+  padding: 0;
+  margin: 12px 0;
+  line-height: 1.6;
+}
+
+a {
+  color: currentColor;
+  text-decoration: none;
+  padding-bottom: 3px;
+  box-shadow: inset 0 -2px 0 0 currentColor;
+}
+
+button {
+  display: block;
+  width: 100%;
+  text-emphasis: center;
+  margin: 0;
+  font-size: 16pt;
+  // color: $dark;
+  background-color: transparent;
+  padding: 6px 0;
+  border-radius: 6px;
+  border: 2px solid transparent;
+  cursor: pointer;
+}
+
+form {
+  padding: 0;
+  margin: 0;
+
+  legend {
+    font-size: 18pt;
+  }
+  fieldset {
+    border: none;
+  }
+}
